+++ tdigest.go
@@ -131,7 +131,7 @@ func weightedAverage(x1 float64, w1 float64, x2 float64, w2 float64) float64 {
 // most common value for this is 1.
 //
 // This will emit an error if `value` is NaN of if `count` is zero.
-func (t *TDigest) AddWeighted(value float64, count uint32) (err error) {
+func (t *TDigest) AddWeighted(value float64, count uint64) (err error) {
 
 	if count == 0 {
 		return fmt.Errorf("Illegal datapoint <value: %.4f, count: %d>", value, count)
@@ -160,7 +160,7 @@ func (t *TDigest) AddWeighted(value float64, count uint32) (err error) {
 	} else {
 		c := float64(t.summary.Count(closest))
 		newMean := weightedAverage(t.summary.Mean(closest), c, value, float64(count))
-		t.summary.setAt(closest, newMean, uint32(c)+count)
+		t.summary.setAt(closest, newMean, uint64(c)+count)
 	}
 	t.count += uint64(count)
 
@@ -218,7 +218,7 @@ func (t *TDigest) Compress() (err error) {
 	t.count = 0
 
 	shuffle(oldTree.means, oldTree.counts, t.rng)
-	oldTree.ForEach(func(mean float64, count uint32) bool {
+	oldTree.ForEach(func(mean float64, count uint64) bool {
 		err = t.AddWeighted(mean, count)
 		return err == nil
 	})
@@ -241,7 +241,7 @@ func (t *TDigest) Merge(other *TDigest) (err error) {
 	data := other.summary.Clone()
 	shuffle(data.means, data.counts, t.rng)
 
-	data.ForEach(func(mean float64, count uint32) bool {
+	data.ForEach(func(mean float64, count uint64) bool {
 		err = t.AddWeighted(mean, count)
 		return err == nil
 	})
@@ -298,7 +298,7 @@ func interpolate(x, x0, x1 float64) float64 {
 //
 // Iteration stops when the supplied function returns false, or when all
 // centroids have been iterated.
-func (t *TDigest) ForEachCentroid(f func(mean float64, count uint32) bool) {
+func (t *TDigest) ForEachCentroid(f func(mean float64, count uint64) bool) {
 	t.summary.ForEach(f)
 }
 
@@ -318,7 +318,7 @@ func (t TDigest) findNeighbors(start int, value float64) (int, int) {
 	return start, lastNeighbor
 }
 
-func (t TDigest) chooseMergeCandidate(begin, end int, value float64, count uint32) int {
+func (t TDigest) chooseMergeCandidate(begin, end int, value float64, count uint64) int {
 	closest := t.summary.Len()
 	sum := t.summary.HeadSum(begin)
 	var n float32
@@ -344,7 +344,7 @@ func (t TDigest) chooseMergeCandidate(begin, end int, value float64, count uint3
 	return closest
 }
 
-func shuffle(means []float64, counts []uint32, rng RNG) {
+func shuffle(means []float64, counts []uint64, rng RNG) {
 	for i := len(means) - 1; i > 1; i-- {
 		j := rng.Intn(i + 1)
 		means[i], means[j], counts[i], counts[j] = means[j], means[i], counts[j], counts[i]
